Move timestamp() to util.[ch]
